# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI CRM Sales Copilot",
    "description": "AI 驅動的房仲業務 CRM 助手",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/leads": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Create Lead",
        "operationId": "create_lead_api_v1_leads_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Leads",
        "operationId": "list_leads_api_v1_leads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "比對姓名或電話",
              "title": "Keyword"
            },
            "description": "比對姓名或電話"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/follow-ups": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Follow Ups",
        "description": "今天該聯絡誰。\n\n路由必須註冊在 /{lead_id} 之前，否則 \"follow-ups\" 會被當成 lead_id\n去比對，然後回一個看起來莫名其妙的 422。",
        "operationId": "list_follow_ups_api_v1_leads_follow_ups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/leads/{lead_id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Get Lead",
        "description": "回傳單一 Lead，並附上互動紀錄 Timeline。",
        "operationId": "get_lead_api_v1_leads__lead_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "leads"
        ],
        "summary": "Update Lead",
        "operationId": "update_lead_api_v1_leads__lead_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "leads"
        ],
        "summary": "Delete Lead",
        "operationId": "delete_lead_api_v1_leads__lead_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/analyze": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Analyze Lead",
        "description": "把客戶原話交給 AI 解析，結果直接寫回這位客戶的需求欄位。\n\n同步等待（約 2～5 秒），前端顯示 loading。\n做成同步是因為 MVP 階段一次只分析一筆，排背景工作要多一個 queue 與輪詢機制，\n複雜度換不到對應的好處；真的要批次分析時再改。\n\n**原話沒變的話不會重新呼叫模型**，直接沿用上一次的結果並把 `reused`\n設成 true。同一段話重算一次結果本來就會一樣，那是純粹的浪費；\n業務改了原話（客戶需求變了）再按，才是真的重新解析。\n\n可能的失敗：\n- 404 這位客戶不存在（或不是你的）\n- 422 這位客戶還沒填原始需求\n- 503 AI 暫時不可用 —— 客戶資料本身不受影響，前端顯示重試按鈕",
        "operationId": "analyze_lead_api_v1_leads__lead_id__analyze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadAnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{lead_id}/follow-up-suggestion": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Suggest Follow Up",
        "description": "產生一則 AI 跟進建議：下一步動作、建議話術、建議時機。\n\n由業務按按鈕觸發，同步等待。不做成「打開待跟進頁面就自動全部產生」——\n一次列表可能有二十位客戶，那就是二十次 API 呼叫，\n而業務今天其實只會打其中三通。\n\n**這支 API 不會改動客戶的任何欄位**，包括下次提醒日。\n建議是參考，不是系統替業務做的決定。\n\n可能的失敗：\n- 404 這位客戶不存在（或不是你的）\n- 422 這位客戶既沒有原始需求也沒有互動紀錄，沒有東西可以據以建議\n- 503 AI 暫時不可用",
        "operationId": "suggest_follow_up_api_v1_leads__lead_id__follow_up_suggestion_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lead_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadFollowUpResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FollowUpResponse": {
        "properties": {
          "viewing_confirm": {
            "items": {
              "$ref": "#/components/schemas/FollowUpItem"
            },
            "type": "array",
            "title": "Viewing Confirm"
          },
          "new_uncontacted": {
            "items": {
              "$ref": "#/components/schemas/FollowUpItem"
            },
            "type": "array",
            "title": "New Uncontacted"
          },
          "due": {
            "items": {
              "$ref": "#/components/schemas/FollowUpItem"
            },
            "type": "array",
            "title": "Due"
          },
          "muted_count": {
            "type": "integer",
            "title": "Muted Count"
          }
        },
        "type": "object",
        "required": [
          "viewing_confirm",
          "new_uncontacted",
          "due",
          "muted_count"
        ],
        "title": "FollowUpResponse",
        "description": "待跟進清單，刻意分成兩堆而不是一份排序好的名單。\n\n「新進未聯絡」與「到期跟進」對應兩種不同的業務動作：\n一個是第一次接觸（搶時間），一個是維繫（別讓它冷掉）。\n混在一起的話，業務打開看到 20 個人，\n分不出哪些是還沒認識、哪些是快跑掉了。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadAnalyzeResponse": {
        "properties": {
          "lead": {
            "$ref": "#/components/schemas/LeadRead"
          },
          "analysis": {
            "$ref": "#/components/schemas/AIAnalysisRead"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "lead",
          "analysis"
        ],
        "title": "LeadAnalyzeResponse",
        "description": "POST /leads/{id}/analyze 的回應。\n\n一併回傳更新後的 lead，前端就不必再打一次 GET —— 少一次來回，\n也少一次「畫面上還是舊資料」的機會。"
      },
      "LeadCreate": {
        "properties": {
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "budget_min": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Min"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 20.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "property_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "building_age_max": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Age Max"
          },
          "parking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parking"
          },
          "purpose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Purpose"
              },
              {
                "type": "null"
              }
            ]
          },
          "purchase_timeline": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Timeline",
            "description": "預計幾個月內購買"
          },
          "urgency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Urgency"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource",
            "default": "OTHER"
          },
          "raw_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Requirement"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "LeadCreate"
      },
      "LeadDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource"
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "raw_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Requirement"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "budget_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Min"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "budget_is_approximate": {
            "type": "boolean",
            "title": "Budget Is Approximate"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "property_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "building_age_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Age Max"
          },
          "parking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parking"
          },
          "purpose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Purpose"
              },
              {
                "type": "null"
              }
            ]
          },
          "purchase_timeline": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Timeline"
          },
          "urgency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Urgency"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_follow_up_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Follow Up At"
          },
          "viewing_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewing Scheduled At"
          },
          "follow_up_muted": {
            "type": "boolean",
            "title": "Follow Up Muted"
          },
          "lead_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Score"
          },
          "lead_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "interactions": {
            "items": {
              "$ref": "#/components/schemas/InteractionRead"
            },
            "type": "array",
            "title": "Interactions",
            "default": []
          },
          "score_reasons": {
            "items": {
              "$ref": "#/components/schemas/ScoreReasonRead"
            },
            "type": "array",
            "title": "Score Reasons",
            "default": []
          },
          "latest_analysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIAnalysisRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_follow_up": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FollowUpAnalysisRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "source",
          "status",
          "raw_requirement",
          "location",
          "budget_min",
          "budget_max",
          "budget_is_approximate",
          "rooms",
          "property_type",
          "building_age_max",
          "parking",
          "purpose",
          "purchase_timeline",
          "urgency",
          "next_follow_up_at",
          "viewing_scheduled_at",
          "follow_up_muted",
          "lead_score",
          "lead_level",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "LeadDetail",
        "description": "Lead Detail 頁專用：在基本資料之外，一併帶出互動紀錄 Timeline。\n\n列表用 LeadRead（輕量），詳細頁用 LeadDetail（含 interactions），\n這樣列表 50 筆客戶時不會順便把幾百筆互動紀錄一起撈出來。"
      },
      "LeadFollowUpResponse": {
        "properties": {
          "suggestion": {
            "$ref": "#/components/schemas/FollowUpAnalysisRead"
          }
        },
        "type": "object",
        "required": [
          "suggestion"
        ],
        "title": "LeadFollowUpResponse",
        "description": "POST /leads/{id}/follow-up-suggestion 的回應。\n\n只回建議，不回 lead —— 因為產生建議不會改到客戶的任何欄位，\n前端手上那份資料仍然是對的，沒必要多傳一份回去。\n這跟 analyze 的回應刻意不一樣，差異本身就在說明兩支 API 的性質不同。"
      },
...
      "LeadListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "LeadListResponse",
        "description": "列表一律包一層，日後要加 total / page 才不用改前端的解析方式。"
      },
      "LeadRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource"
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "raw_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Requirement"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "budget_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Min"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "budget_is_approximate": {
            "type": "boolean",
            "title": "Budget Is Approximate"
          },
          "rooms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rooms"
          },
          "property_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PropertyType"
              },
              {
                "type": "null"
              }
            ]
          },
          "building_age_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Building Age Max"
          },
          "parking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parking"
          },
          "purpose": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Purpose"
              },
              {
                "type": "null"
              }
            ]
          },
          "purchase_timeline": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchase Timeline"
          },
          "urgency": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Urgency"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_follow_up_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Follow Up At"
          },
          "viewing_scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Viewing Scheduled At"
          },
          "follow_up_muted": {
            "type": "boolean",
            "title": "Follow Up Muted"
          },
          "lead_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lead Score"
          },
          "lead_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "source",
          "status",
          "raw_requirement",
          "location",
          "budget_min",
          "budget_max",
          "budget_is_approximate",
          "rooms",
          "property_type",
          "building_age_max",
          "parking",
          "purpose",
          "purchase_timeline",
          "urgency",
          "next_follow_up_at",
          "viewing_scheduled_at",
          "follow_up_muted",
          "lead_score",
          "lead_level",
          "owner_id",
          "created_at",
          "updated_at"
        ],
        "title": "LeadRead",
        "description": "回傳給前端的完整 Lead。"
      },
...
      "LeadStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "CONTACTED",
          "INTERESTED",
          "MEETING",
          "VIEWING",
          "NEGOTIATING",
          "WON",
          "LOST"
        ],
        "title": "LeadStatus",
        "description": "Lead 在銷售漏斗中的位置（對應 Dashboard 的 Lead Funnel）。"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "貼上登入取得的 access_token",
        "scheme": "bearer"
      }
    }
  }
}